requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
anyio==4.3.0
apache-airflow==2.9.0
apache-airflow-providers-celery==3.6.1
apache-airflow-providers-common-io==1.3.0
apache-airflow-providers-common-sql==1.11.1
apache-airflow-providers-fab==1.0.2
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.10.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.0
appdirs==1.4.4
argcomplete==3.2.3
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
boto3==1.34.93
botocore==1.34.93
cachelib==0.9.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.0.0
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
commonmark==0.9.1
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.2.1
cron-descriptor==1.4.3
croniter==2.0.3
cryptography==41.0.7
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.25
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docker==7.0.0
docutils==0.16
email-validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Login==0.6.3
flask-session==0.5.0
Flask-SQLAlchemy==2.5.1
flask-wtf==1.2.1
flower==2.0.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.12.2
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.43
google-re2==1.1
googleapis-common-protos==1.63.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.22.2
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.4.0
inflection==0.5.1
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
kaggle==1.6.12
kiwisolver==1.4.5
kombu==5.3.6
lazy-object-proxy==1.10.0
lightning==2.2.3
lightning-utilities==0.11.2
limits==3.10.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mlflow==2.12.1
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.0.0
pendulum==3.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.1
pluggy==1.4.0
prison==0.2.1
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.22
pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.15.0
python-slugify==8.0.4
pytorch-lightning==2.2.3
pytorch-lightning-bolts==0.2.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
querystring-parser==1.2.4
referencing==0.31.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.1.0
rich-argparse==1.4.0
rpds-py==0.18.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tasklib==2.5.1
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.4.0
time-machine==2.14.1
tokenizers==0.19.1
torch==2.3.0+cu118
torchmetrics==1.3.2
torchvision==0.18.0+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
treelib==1.6.4
triton==2.3.0
typing-extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.2.2
urllib3==1.26.18
urwid==2.6.11
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.16.0
wtforms==3.1.2
yarl==1.9.4
zipp==3.18.1
